fix small error

+++ src/FullPipe/InferNode.cpp
@@ -36,7 +36,7 @@ void InferNodeWorker::process(std::size_t batchIdx)
 
             auto ptrVideoBuf = m_vecBlobInput[0]->get<int, VideoMeta>(0);
 
-            uint64_t fd = *((uint64_t *)ptrVideoBuf->getPtr());
+            uint64_t fd = *((int32_t *)ptrVideoBuf->getPtr());
 
             VideoMeta *ptrVideoMeta = ptrVideoBuf->getMeta();
             int input_height = ptrVideoMeta->videoHeight;
@@ -113,7 +113,7 @@ void InferNodeWorker::process(std::size_t batchIdx)
             auto ptrBufInfer = m_vecBlobInput[0]->get<int, InferMeta>(0);
             auto ptrBufVideo = m_vecBlobInput[0]->get<int, VideoMeta>(1);
 
-            uint64_t fd = *((uint64_t *)ptrBufVideo->getPtr());
+            uint64_t fd = *((int32_t *)ptrBufVideo->getPtr());
 
             VideoMeta *ptrVideoMeta = ptrBufVideo->getMeta();
             int input_height = ptrVideoMeta->videoHeight;
